Refactor: Extract TranscriptMixin (PR 1)
